requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
slack_bolt>=1.18.0
google-generativeai>=0.4.0
openai>=1.12.0
ollama>=0.1.6
anthropic>=0.18.0
mistralai>=0.1.0
groq>=0.4.0
dashscope>=1.14.0
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
watchdog>=4.0.0
